gpu: nvgpu: Deterministic submit fix
Fix some simple pointer arithmetic errors in the deterministic submit
path. The lockless allocator was doing a subtraction to determine the
correct offset of the element to free. However, this subtraction was
using the base address and a numeric offset - not a pointer offset. Thus
the difference computed was in bytes not in elements of the block size.

The fix is simple: just divide by the block size.

Also this modifies the debugging statement a bit so that a bit more
information is printed at more useful times.

Lastly, a pointer to numeric cast was fixed in the fence code.

Reviewed-on: https://git-master.nvidia.com/r/1538905
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/gk20a/fence_gk20a.c
@@ -51,7 +51,7 @@ static void gk20a_fence_free(struct kref *ref)
 
 	if (f->allocator) {
 		if (nvgpu_alloc_initialized(f->allocator))
-			nvgpu_free(f->allocator, (size_t)f);
+			nvgpu_free(f->allocator, (u64)(uintptr_t)f);
 	} else
 		nvgpu_kfree(g, f);
 }
